chore: move network test to the right package
