+++ docs/features/networking.md
@@ -129,7 +129,7 @@ But according to those docs, it's supported only for Linux hosts:
 In the case you need to skip a test on non-Linux hosts, you can use the `SkipIfDockerDesktop` function:
 
 <!--codeinclude-->
-[Skipping tests on non-Linux hosts](../../docker_test.go) inside_block:skipIfDockerDesktop
+[Skipping tests on non-Linux hosts](../../network/network_test.go) inside_block:skipIfDockerDesktop
 <!--/codeinclude-->
 
 It will try to get a Docker client and obtain its Info. In the case the Operating System is "Docker Desktop", it will skip the test.
