+++ network/network_test.go
@@ -2,9 +2,11 @@ package network_test
 
 import (
 	"context"
+	"os"
 	"testing"
 	"time"
 
+	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/filters"
 	dockernetwork "github.com/docker/docker/api/types/network"
 	"github.com/stretchr/testify/require"
@@ -482,4 +484,28 @@ func TestWithNewNetworkContextTimeout(t *testing.T) {
 func TestCleanupWithNil(t *testing.T) {
 	var network *testcontainers.DockerNetwork
 	testcontainers.CleanupNetwork(t, network)
+}
+
+func TestContainerWithNetworkModeAndNetworkTogether(t *testing.T) {
+	if os.Getenv("XDG_RUNTIME_DIR") != "" {
+		t.Skip("Skipping test that requires host network access when running in a container")
+	}
+
+	// skipIfDockerDesktop {
+	ctx := context.Background()
+	testcontainers.SkipIfDockerDesktop(t, ctx)
+	// }
+
+	nginx, err := testcontainers.Run(
+		ctx, nginxAlpineImage,
+		testcontainers.WithHostConfigModifier(func(hc *container.HostConfig) {
+			hc.NetworkMode = "host"
+		}),
+		network.WithNetworkName([]string{"nginx"}, "new-network"),
+	)
+	testcontainers.CleanupContainer(t, nginx)
+	if err != nil {
+		// Error when NetworkMode = host and Network = []string{"bridge"}
+		t.Logf("Can't use Network and NetworkMode together, %s\n", err)
+	}
 }
